added documentation to functions and moved image-api into user-api
