+++ App.ts
@@ -1,7 +1,8 @@
-import { App } from './src';
-// import StorybookUIRoot from './.storybook';
+// import { App } from './src';
+import StorybookUIRoot from './.storybook';
 
 // const STORYBOOK_START = true;
 // console.log(process.env.STORYBOOK_START);
 // export default process.env.STORYBOOK_START ? StorybookUIRoot : App;
-export default App;
+// export default App;
+export default StorybookUIRoot;
